Move admin hooks to admin folder
